json完善

+++ service/metro.go
@@ -5,6 +5,6 @@ import (
 	"metro/model"
 )
 
-func AddMetro(metro *model.Metro) (id int) {
+func AddMetro(metro *model.Metro) (id int64) {
 	return dao.InsertMetros(metro)
 }
